+++ src/app.ts
@@ -5,72 +5,10 @@
 /// <reference path="decorators/autobind.ts" />
 /// <reference path="components/base-component.ts" />
 /// <reference path="components/project-item.ts" />
+/// <reference path="components/project-list.ts" />
 
 namespace App {
-  // ProjectList Class
-  class ProjectList extends Component<HTMLDivElement, HTMLElement> implements DragTarget {
-    assignedProjects: Project[];
 
-    constructor(private type: 'active' | 'finished') {
-      super("project-list", "app", false, `${type}-projects`);
-      this.assignedProjects = [];
-      this.configure();
-      this.renderContent();
-    }
-
-    configure() {
-      projectState.addListener((projects: Project[]) => {
-        this.assignedProjects = projects.filter((project: Project) => {
-          if (this.type === "active") return project.status === ProjectStatus.Active;
-          return project.status === ProjectStatus.Finished;
-        });
-        this.renderProjects();
-      });
-      this.element.addEventListener('dragover', this.dragOverHandler);
-      this.element.addEventListener('dragleave', this.dragLeaveHandler);
-      this.element.addEventListener('drop', this.dropHandler);
-    }
-
-    renderContent() {
-      const listId = `${this.type}-projects-list`;
-      this.element.querySelector('ul')!.id = listId;
-      this.element.querySelector('h2')!.textContent = this.type.toUpperCase() + ' PROJECTS';
-    }
-
-    @autobind
-    dragOverHandler(event: DragEvent) {
-      if (event.dataTransfer && event.dataTransfer.types[0] === "text/plain") {
-        event.preventDefault();
-        const listEl = this.element.querySelector("ul")!;
-        listEl.classList.add('droppable');
-      }
-    }
-
-    @autobind
-    dragLeaveHandler(_: DragEvent) {
-      const listEl = this.element.querySelector("ul")!;
-      listEl.classList.remove('droppable');
-    }
-
-    @autobind
-    dropHandler(event: DragEvent) {
-      event.preventDefault();
-      const projectId = event.dataTransfer!.getData("text/plain");
-      projectState.moveProject(
-        projectId,
-        this.type === 'active' ? ProjectStatus.Active : ProjectStatus.Finished
-      );
-    }
-
-    private renderProjects() {
-      const listEl = document.getElementById(`${this.type}-projects-list`)! as HTMLUListElement;
-      listEl.innerHTML = '';
-      for (const projectItem of this.assignedProjects) {
-        new ProjectItem(this.element.querySelector('ul')!.id, projectItem);
-      }
-    }
-
-  }
 
   // ProjectInput Class
   class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
